creates new ServerUtils file and removed image from uploads folder
